Move type inference types to separate file and re-enable linting
